Prevent rbenv-binstubs from shadowing system commands from PATH
Previously if a binstub was created by an arbitrary ruby package
anywhere, the generated binstub will be in effect everywhere else. This
caused projects that generated common binstub names such as `watch` to
"shadow" your system paths and break stuff.

This adds a filter to the which command to prevent that from happening.
If the binstub path does not exist, then the system path is consulted.